requirements.txt
asyncio>=3.4.3
pathlib>=1.0.1
playwright>=1.48.0
selenium>=4.25.0
selenium-wire>=5.1.0
trafilatura>=1.12.2
webdriver-manager>=4.0.2
requests>=2.32.3
beautifulsoup4>=4.12.3
python-dotenv
click>=8.1.7
google-cloud-storage>=2.18.2
undetected-chromedriver>=3.5.5
yagmail
uritemplate
oauthlib
httplib2
requests-oauthlib
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
bs4